fixed and tidied up instrument router tests
